prod upd 13.12.2022

+++ src/app/components/common/filters-components/components/components/filter-menu/filter-menu.component.ts
@@ -2,7 +2,8 @@
 import { AfterViewInit, Component, ElementRef, Input, OnDestroy, ViewChild } from '@angular/core';
 import { FiltersService } from '@app/components/common/filters-components/services/filters.service';
 import { TagsService } from '@app/components/common/filters-components/services/tags.service';
-import { Subscription } from 'rxjs';
+import { GamesService } from '@app/core/services/assets/games.service';
+import { concatMap, Subscription } from 'rxjs';
 
 @Component({
   selector: 'filter-menu',
@@ -14,11 +15,14 @@ export class FilterMenuComponent implements AfterViewInit, OnDestroy {
   constructor(
     private tagsService: TagsService,
     private filtersService: FiltersService,
+    private gamesService: GamesService,
   ) { }
 
   menuActive = false;
   checkedItems: any = [];
+  loadingItems: boolean | null = false;
   sub!: Subscription;
+  page = 1;
 
   @ViewChild('wrapper') wrapper!: ElementRef;
   @ViewChild('inputContainer') inputContainer!: ElementRef;
@@ -26,13 +30,15 @@ export class FilterMenuComponent implements AfterViewInit, OnDestroy {
 
   @Input() inputType = 'checkbox';
   @Input() title = 'Title'
+  @Input() type = '';
   @Input() menuHeight = '200px';
   @Input() searchType: string | null = 'search';
-  @Input() items: any[] = [{ name: 'Game', genre: 'casual', selected: false }, { name: 'Game', genre: 'casual', selected: true }, { name: 'Game', genre: 'casual', selected: true }, { name: 'Game', genre: 'casual', selected: true }, { name: 'Game', genre: 'casual', selected: true }, { name: 'Game', genre: 'casual', selected: true },];
+  @Input() items!: any[];
 
 
   ngOnInit() {
     this.resetFilters$();
+    this.processMenuContent();
   }
 
   ngAfterViewInit() {
@@ -71,8 +77,9 @@ export class FilterMenuComponent implements AfterViewInit, OnDestroy {
   onChangeInput(event: any) {
     const value = event.target.value;
     const reference = event.target;
+    console.log('radio value', event.target.value);
     if (this.inputType === 'radio') {
-      this.tagsService.removeTag(this.checkedItems[0]);
+      if(this.checkedItems[0]) this.tagsService.removeTag(this.checkedItems[0]);
       this.checkedItems = [{ value: value, type: this.title, reference: reference }];
       this.tagsService.addTag = { value: value, type: this.title, inputType: this.inputType, reference: reference };
     }
@@ -95,8 +102,35 @@ export class FilterMenuComponent implements AfterViewInit, OnDestroy {
     }
   }
 
+  processMenuContent() {
+    if (this.type == 'games') {
+        this.loadMoreGames();
+    }
+  }
+
   onInput(e: any) {
     e.target.checked = false;
+  }
+
+  loadMoreGames() {
+    if(this.loadingItems == true && this.type =='games') return;
+    this.loadingItems = true;
+    this.gamesService.fetchGames(this.page)
+        .subscribe(games => {
+          if (!this.items) this.items = [];
+
+          for (let game of games) this.items.push(game);
+
+          if(games?.length < 10) {
+            this.loadingItems = null;
+            return;
+          }
+          this.loadingItems = false;
+          this.page++;
+        });
+  }
+  scrolledToBottom() {
+    this.loadMoreGames();
   }
 
   ngOnDestroy(): void {
